+++ editor/package.json
@@ -26,10 +26,13 @@
     "@formatjs/intl-localematcher": "^0.6.1",
     "@grida/bitmap": "workspace:*",
     "@grida/cmath": "workspace:*",
+    "@grida/io-figma": "workspace:*",
     "@grida/pixel-grid": "workspace:*",
     "@grida/ruler": "workspace:*",
+    "@grida/io-svg": "workspace:*",
     "@grida/tokens": "workspace:*",
     "@grida/transparency-grid": "workspace:*",
+    "@grida/vn": "workspace:*",
     "@hookform/resolvers": "^4.1.3",
     "@mdx-js/loader": "^3.0.1",
     "@mdx-js/react": "^3.0.1",
@@ -137,7 +140,6 @@
     "html-to-image": "^1.11.11",
     "i18next": "^23.7.9",
     "immer": "^9.0.3",
-    "inline-style-parser": "^0.2.4",
     "input-otp": "^1.4.2",
     "libphonenumber-js": "^1.12.6",
     "lowlight": "^3.3.0",
@@ -184,8 +186,6 @@
     "signature_pad": "^4.2.0",
     "stylis": "^4.3.2",
     "svg-pathdata": "^7.1.0",
-    "svgo": "^3.3.2",
-    "svgson": "^5.3.1",
     "swr": "^2.2.5",
     "tailwind-merge": "^2.2.2",
     "timezones-list": "^3.0.3",
