+++ packages/grida-canvas-io-figma/index.ts
@@ -1,0 +1,2 @@
+import { iofigma } from "./lib";
+export default iofigma;
