+++ packages/grida-canvas-io-svg/index.ts
@@ -1,0 +1,2 @@
+import { iosvg } from "./lib";
+export default iosvg;
