+++ packages/grida-canvas-io-svg/svgson.example.json
