+++ packages/grida-canvas-vn/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@grida/vn",
+  "description": "Vector Network Spec",
+  "scripts": {
+    "test": "jest"
+  },
+  "dependencies": {
+    "@grida/cmath": "workspace:*",
+    "svg-pathdata": "^7.1.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.2",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.5"
+  },
+  "jest": {
+    "preset": "ts-jest"
+  }
+}
